create mixin for pagination in models
